Trait sui_synthetic_ingestion::benchmark::BenchmarkableIndexer

pub trait BenchmarkableIndexer {
    // Required methods
    fn subscribe_to_committed_checkpoints(
        &self,
    ) -> Receiver<Option<IndexerProgress>>;
    fn start<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn stop<'async_trait>(
        self,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait;
}

A trait that can be implemented on top of any indexer to benchmark its throughput. It will generate synthetic transactions and checkpoints as ingestion source.

Required Methods

fn subscribe_to_committed_checkpoints( &self, ) -> Receiver<Option<IndexerProgress>>

Allows the benchmark to subscribe and monitor the committed checkpoints progress. This is needed both in order to log periodic throughput, but also to know when the benchmark can stop.

fn start<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Start the indexer. Note that we only start a timer before calling this function. So the implementation should only start the indexer when this function is called.

fn stop<'async_trait>( self, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait,

Stop the indexer. This would allow the benchmark to exit.